backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3

# AI/ML
anthropic>=0.8.1
qdrant-client>=1.7.1

# Database
supabase>=2.3.0
redis>=5.0.1

# Utils
python-dotenv>=1.0.0
httpx>=0.26.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
